+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.dto;
+package kitchenpos.product.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductResponse {
     private Long id;
